+++ package.json
@@ -9,26 +9,28 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@chakra-ui/icons": "^2.0.10",
     "@chakra-ui/next-js": "^2.1.5",
     "@chakra-ui/react": "^2.8.0",
-    "@chakra-ui/icons": "^2.0.10",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@splidejs/react-splide": "^0.7.12",
     "@types/node": "20.5.7",
     "@types/react": "18.2.21",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.15",
+    "cheerio": "^1.0.0-rc.12",
     "eslint": "8.48.0",
     "eslint-config-next": "13.4.19",
     "framer-motion": "^10.16.1",
+    "marked": "^9.0.0",
     "next": "13.4.19",
     "postcss": "8.4.28",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwindcss": "3.3.3",
-    "typescript": "5.2.2",
+    "react-helmet": "^6.1.0",
     "react-icons": "^4.4.0",
-    "@splidejs/react-splide": "^0.7.12",
-    "react-helmet": "^6.1.0"
+    "tailwindcss": "3.3.3",
+    "typescript": "5.2.2"
   }
 }
